pyproject.toml
[project]
name = "steering_agent"
version = "0.1"
description = "AI-driven agent designed to facilitate aerospace requirements validation"
authors = []
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
google-adk = "^1.3.0"
google-genai = "^1.20.0"
pydantic = "^2.10.6"
python-dotenv = "^1.0.1"
google-cloud-aiplatform = { version = "^1.97.0", extras = [
    "adk",
    "agent-engines",
] }
reportlab = "^4.4.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
google-adk = { extras = ["eval"], version = "^1.0.0" }
google-cloud-aiplatform = { version = "^1.97.0", extras = [
    "adk",
    "agent-engines",
    "evaluation",
] }
pytest = "^8.3.5"
black = "^25.1.0"
pytest-asyncio = "^0.26.0"
pandas = "^2.2.3"
tabulate = "^0.9.0"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.3.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"